refactor(agent-sdk-ts): reorganize sdk layout
