+++ packages/agent-sdk-ts/src/browser.ts
@@ -1,3 +1,3 @@
 // Browser-targeted entry that exposes only pure data models and type guards
 // to avoid bundling Node-specific workspace and tool implementations.
-export * from './types';
+export * from './sdk/types';
