+++ packages/agent-sdk-ts/src/index.ts
@@ -1,6 +1,3 @@
-export * from './types';
-export * from './workspace';
+export * from './sdk';
 export * from './tools';
-export * from './runtime';
-export * from './llm';
-export * from './conversation';
+export * from './workspace';
