+++ packages/agent-sdk-ts/src/sdk/llm/openai-compatible.ts
