+++ packages/agent-sdk-ts/src/sdk/runtime/StuckDetector.ts
