+++ packages/agent-sdk-ts/tsconfig.json
@@ -16,5 +16,5 @@
     "skipLibCheck": true
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["src/__tests__/**"]
+  "exclude": ["src/**/__tests__/**"]
 }
